fix: rustfmt + clippy - abs_diff, unnecessary u128 casts, formatting

+++ crates/los-network/src/fee_scaling.rs
@@ -273,7 +273,7 @@ impl BlockBurnState {
     pub fn get_capacity_percentage_bps(&self) -> u32 {
         let used = BURN_LIMIT_PER_BLOCK_CIL - self.remaining_capacity;
         // (used * 10000) / total gives basis points (0.01% precision)
-        ((used as u128 * 10_000) / BURN_LIMIT_PER_BLOCK_CIL as u128) as u32
+        ((used * 10_000) / BURN_LIMIT_PER_BLOCK_CIL) as u32
     }
 
     /// Get fill percentage (0-100) — DEPRECATED: use get_capacity_percentage_bps() instead
